♻️ refactor: [STRUCTURAL] SSE スタブを共有テストヘルパーに切り出す
useChatStream.test.tsx が持っていた streamingFetchStub / tokenEvent /
doneEvent を src/test/streamingFetchStub.ts へ移し、他のテストからも
チャット SSE を手で流し込めるようにする。
